split core to break import cycle
