+++ core/configs/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["configs.go"],
+    importpath = "github.com/bazelbuild/bazelisk/core/configs",
+    visibility = ["//visibility:public"],
+    deps = ["//core/path:go_default_library"],
+)
