+++ core/path/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["path.go"],
+    importpath = "github.com/bazelbuild/bazelisk/core/path",
+    visibility = ["//visibility:public"],
+)
