+++ platforms/BUILD
@@ -6,12 +6,19 @@ go_library(
     importpath = "github.com/bazelbuild/bazelisk/platforms",
     visibility = ["//visibility:public"],
     deps = [
+        "//core/configs:go_default_library",
         "@com_github_hashicorp_go_version//:go_default_library",
     ],
 )
 
 go_test(
     name = "platforms_test",
+    srcs = ["platforms_test.go"],
+    embed = [":go_default_library"],
+)
+
+go_test(
+    name = "go_default_test",
     srcs = ["platforms_test.go"],
     embed = [":go_default_library"],
 )
